# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigitPro Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  coinBalance: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface CoinBalance {
  balance: number;
  currency: string;
}

export interface CoinPackage {
  id: string;
  coins: number;
  priceXOF: number;
  priceUSD: number;
  popular: boolean;
  bonus?: number;
}

export type TopupRequestPaymentMethod =
  (typeof TopupRequestPaymentMethod)[keyof typeof TopupRequestPaymentMethod];

export const TopupRequestPaymentMethod = {
  crypto: "crypto",
  mtn_momo: "mtn_momo",
  moov_money: "moov_money",
} as const;

export interface TopupRequest {
  packageId: string;
  paymentMethod: TopupRequestPaymentMethod;
}

export interface TopupResponse {
  paymentUrl: string;
  orderId: string;
  amount: number;
  currency: string;
}

export type CoinTransactionType =
  (typeof CoinTransactionType)[keyof typeof CoinTransactionType];

export const CoinTransactionType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface CoinTransaction {
  id: string;
  type: CoinTransactionType;
  amount: number;
  description: string;
  service?: string;
  createdAt: string;
}

export interface WebTemplate {
  id: string;
  name: string;
  category: string;
  description: string;
  previewImage: string;
  coinCost: number;
  tags: string[];
  featured: boolean;
}

export type CreateWebsiteRequestCustomizations = { [key: string]: unknown };

export interface CreateWebsiteRequest {
  templateId: string;
  siteName: string;
  tagline?: string;
  logoUrl?: string;
  primaryColor?: string;
  customizations?: CreateWebsiteRequestCustomizations;
}

export type WebsiteStatus = (typeof WebsiteStatus)[keyof typeof WebsiteStatus];

export const WebsiteStatus = {
  active: "active",
  inactive: "inactive",
  pending: "pending",
} as const;

export interface Website {
  id: string;
  siteName: string;
  templateId: string;
  templateName: string;
  url: string;
  status: WebsiteStatus;
  createdAt: string;
}

export interface PhoneNumber {
  id: string;
  number: string;
  country: string;
  countryCode: string;
  services: string[];
  coinCost: number;
  durationDays: number;
}

export interface PurchasePhoneRequest {
  phoneId: string;
  service: string;
}

export type PurchasedPhoneStatus =
  (typeof PurchasedPhoneStatus)[keyof typeof PurchasedPhoneStatus];

export const PurchasedPhoneStatus = {
  active: "active",
  expired: "expired",
  used: "used",
} as const;

export interface PurchasedPhone {
  id: string;
  number: string;
  country: string;
  service: string;
  expiresAt: string;
  smsCode?: string;
  status: PurchasedPhoneStatus;
  createdAt: string;
}

export type CreateTransferRequestType =
  (typeof CreateTransferRequestType)[keyof typeof CreateTransferRequestType];

export const CreateTransferRequestType = {
  simulation: "simulation",
  document: "document",
} as const;

export interface CreateTransferRequest {
  type: CreateTransferRequestType;
  senderName: string;
  senderIban?: string;
  senderBank: string;
  receiverName: string;
  receiverIban?: string;
  receiverBank: string;
  amount: number;
  currency: string;
  reference?: string;
  description?: string;
  transferDate?: string;
  estimatedArrival?: string;
}

export type TransferType = (typeof TransferType)[keyof typeof TransferType];

export const TransferType = {
  simulation: "simulation",
  document: "document",
} as const;

export type TransferStatus =
  (typeof TransferStatus)[keyof typeof TransferStatus];

export const TransferStatus = {
  initiated: "initiated",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Transfer {
  id: string;
  type: TransferType;
  trackingCode: string;
  trackingUrl: string;
  senderName: string;
  senderIban?: string;
  senderBank: string;
  receiverName: string;
  receiverIban?: string;
  receiverBank: string;
  amount: number;
  currency: string;
  reference?: string;
  description?: string;
  status: TransferStatus;
  progress: number;
  transferDate?: string;
  estimatedArrival?: string;
  coinCost: number;
  createdAt: string;
}

export type TransferTrackingStatus =
  (typeof TransferTrackingStatus)[keyof typeof TransferTrackingStatus];

export const TransferTrackingStatus = {
  initiated: "initiated",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface TransferStep {
  label: string;
  description: string;
  completed: boolean;
  timestamp?: string;
}

export interface TransferTracking {
  trackingCode: string;
  senderName: string;
  senderBank: string;
  receiverName: string;
  receiverBank: string;
  amount: number;
  currency: string;
  status: TransferTrackingStatus;
  progress: number;
  transferDate?: string;
  estimatedArrival?: string;
  steps: TransferStep[];
}

export type GetTemplatesParams = {
  category?: string;
};

export type GetAvailablePhonesParams = {
  service?: string;
  country?: string;
};