# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DigitPro Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /coins/topup:
    post:
      operationId: topupCoins
      tags: [coins]
      summary: Initiate coin top-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupRequest"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupResponse"
...
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: Get website templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebTemplate"
  /templates/create:
    post:
      operationId: createWebsite
      tags: [templates]
      summary: Create a website from a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebsiteRequest"
      responses:
        "201":
          description: Website created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
...
  /phones/available:
    get:
      operationId: getAvailablePhones
      tags: [phones]
      summary: Get available phone numbers
      parameters:
        - name: service
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Available numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PhoneNumber"
  /phones/purchase:
    post:
      operationId: purchasePhone
      tags: [phones]
      summary: Purchase a phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasePhoneRequest"
      responses:
        "201":
          description: Phone purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasedPhone"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
      required:
        - email
        - password
        - firstName
        - lastName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        coinBalance:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - coinBalance
        - createdAt
...
    TopupRequest:
      type: object
      properties:
        packageId:
          type: string
        paymentMethod:
          type: string
          enum: [crypto, mtn_momo, moov_money]
      required:
        - packageId
        - paymentMethod
    TopupResponse:
      type: object
      properties:
        paymentUrl:
          type: string
        orderId:
          type: string
        amount:
          type: number
        currency:
          type: string
      required:
        - paymentUrl
        - orderId
        - amount
        - currency
...
    WebTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        previewImage:
          type: string
        coinCost:
          type: integer
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
      required:
        - id
        - name
        - category
        - description
        - previewImage
        - coinCost
        - tags
        - featured
    CreateWebsiteRequest:
      type: object
      properties:
        templateId:
          type: string
        siteName:
          type: string
        tagline:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        customizations:
          type: object
          additionalProperties: true
      required:
        - templateId
        - siteName
    Website:
      type: object
      properties:
        id:
          type: string
        siteName:
          type: string
        templateId:
          type: string
        templateName:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - siteName
        - templateId
        - templateName
        - url
        - status
        - createdAt
    PhoneNumber:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        country:
          type: string
        countryCode:
          type: string
        services:
          type: array
          items:
            type: string
        coinCost:
          type: integer
        durationDays:
          type: integer
      required:
        - id
        - number
        - country
        - countryCode
        - services
        - coinCost
        - durationDays
    PurchasePhoneRequest:
      type: object
      properties:
        phoneId:
          type: string
        service:
          type: string
      required:
        - phoneId
        - service
    PurchasedPhone:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        country:
          type: string
        service:
          type: string
        expiresAt:
          type: string
          format: date-time
        smsCode:
          type: string
        status:
          type: string
          enum: [active, expired, used]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - number
        - country
        - service
        - expiresAt
        - status
        - createdAt
...