# lib/api-zod/src/generated/types/phoneNumber.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigitPro Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PhoneNumber {
  id: string;
  number: string;
  country: string;
  countryCode: string;
  services: string[];
  coinCost: number;
  durationDays: number;
}