# lib/api-zod/src/generated/types/topupResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigitPro Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TopupResponse {
  paymentUrl: string;
  orderId: string;
  amount: number;
  currency: string;
}