# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigitPro Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  coinBalance: number;
  createdAt: Date;
}